Moving RIPE and HELMET to apps
