+++ idaes/apps/ripe/atermconstruct.py
@@ -11,7 +11,7 @@
 # license information.
 #################################################################################
 import numpy as np
-from idaes.surrogate.ripe import mechs
+from idaes.apps.ripe import mechs
 from .shared import sharedata as sd
 # from ripe.shared import debug as debug
