+++ idaes/apps/ripe/kinforms.py
