+++ idaes/apps/ripe/shared.py
